container-disk: new implementation using pidfile
The new container disk binary write its own file into a known location,
in this way virt-launcher is able to locate the container disk
artifacts by looking into the proc filesystem for the pid.

Bump version to v3alpha.
