--- cmd/container-disk-v2alpha/container_disk_v2alpha_suite_test.go
@@ -1,11 +1,0 @@
-package container_disk_v2alpha_test
-
-import (
-	"kubevirt.io/client-go/testutils"
-
-	"testing"
-)
-
-func TestContainerDiskV2alpha(t *testing.T) {
-	testutils.KubeVirtTestSuiteSetup(t)
-}
