+++ cmd/virt-launcher/BUILD.bazel
@@ -58,7 +58,7 @@ pkg_tar(
     srcs = [
         "node-labeller/node-labeller.sh",
         ":virt-launcher",
-        "//cmd/container-disk-v2alpha:container-disk",
+        "//cmd/container-disk-v3alpha:container-disk",
         "//cmd/virt-freezer",
         "//cmd/virt-launcher-monitor",
         "//cmd/virt-probe",
